+++ app/Projects/Project.php
@@ -9,18 +9,17 @@ use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Support\Carbon;
 
 /**
- * A Project is a container for one or more Applications or Redirects.
+ * A Project is a container for one or more ProjectServices.
  *
- * @property int                           $id
- * @property string                        $name
- * @property bool                          $is_enabled
- * @property ?Carbon                       $created_at
- * @property ?Carbon                       $updated_at
- * @property array<Application>|Collection $applications
- * @property ?int                          $applications_count
- * @property array<Redirect>|Collection    $redirects
- * @property ?int                          $redirects_count
+ * @property int                              $id
+ * @property string                           $name
+ * @property bool                             $is_enabled
+ * @property ?Carbon                          $created_at
+ * @property ?Carbon                          $updated_at
+ * @property array<ProjectService>|Collection $services
+ * @property ?int                             $services_count
  *
+ * @method static Project         create()
  * @method static Builder|Project query()
  * @method static Builder|Project whereCreatedAt($value)
  * @method static Builder|Project whereId($value)
@@ -45,13 +44,8 @@ class Project extends Model
 
     protected $table = 'projects';
 
-    public function applications(): HasMany
+    public function services(): HasMany
     {
-        return $this->hasMany(Application::class);
-    }
-
-    public function redirects(): HasMany
-    {
-        return $this->hasMany(Redirect::class);
+        return $this->hasMany(ProjectService::class);
     }
 }