+++ app/Projects/RequiresNginxData.php
@@ -2,8 +2,8 @@
 
 namespace Servidor\Projects;
 
+use Illuminate\Support\Collection;
 use Servidor\Projects\Actions\MissingProjectData;
-use Servidor\Projects\Applications\Templates\Template;
 
 /**
  * @property array<string, string> $requiredNginxData
@@ -12,19 +12,38 @@ trait RequiresNginxData
 {
     public function checkNginxData(): void
     {
-        $type = $this instanceof Redirect ? 'Redirect' : 'App';
-
         foreach ($this->requiredNginxData as $property => $name) {
-            if ($this instanceof Template && $this->getApp()->{$property}) {
-                continue;
-            }
-            if ($this instanceof Redirect && $this->{$property}) {
+            if ($this->checkConfig($property)) {
                 continue;
             }
 
             throw new MissingProjectData(
-                sprintf('%s does not have a %s set.', $type, $name),
+                sprintf('Service does not have a %s set.', $name),
             );
         }
+    }
+
+    /** @phan-suppress PhanUndeclaredMethod, PhanUndeclaredProperty */
+    private function checkConfig(string $property): bool
+    {
+        $data = $this->getService();
+
+        if ('config.' !== mb_substr($property, 0, 7)) {
+            return isset($data->{$property});
+        }
+
+        $config = $data->config ?: new Collection();
+        $property = mb_substr($property, 7);
+
+        if (str_contains($property, '.')) {
+            [$key, $property] = explode('.', $property);
+
+            /** @var array $key */
+            $key = $config->get($key, []);
+
+            return isset($key[$property]);
+        }
+
+        return $config->has($property);
     }
 }