+++ tests/Feature/Api/Projects/Actions/SaveSslCertTest.php
@@ -3,9 +3,8 @@
 namespace Tests\Feature\Api\Projects\Actions;
 
 use Illuminate\Foundation\Testing\RefreshDatabase;
-use Servidor\Projects\Application;
 use Servidor\Projects\Project;
-use Servidor\Projects\Redirect;
+use Servidor\Projects\ProjectService;
 use Tests\TestCase;
 
 class SaveSslCertTest extends TestCase
@@ -22,7 +21,7 @@ class SaveSslCertTest extends TestCase
         $this->assertFileDoesNotExist($key);
 
         $project = Project::create(['name' => 'SSL Test']);
-        $project->applications()->save($app = new Application([
+        $project->services()->save($service = new ProjectService([
             'domain_name' => 'ssl.test',
             'config' => [
                 'ssl' => true,
@@ -45,11 +44,14 @@ class SaveSslCertTest extends TestCase
         $this->assertFileDoesNotExist($key);
 
         $project = Project::create(['name' => 'SSL RTest']);
-        $project->redirects()->save($redirect = new Redirect([
+        $project->services()->save($service = new ProjectService([
             'domain_name' => 'ssl.rtest',
-            'target' => 'foo',
-            'type' => 301,
+            'template' => 'redirect',
             'config' => [
+                'redirect' => [
+                    'target' => 'foo',
+                    'type' => 301,
+                ],
                 'ssl' => true,
                 'sslCertificate' => 'redirect cert',
                 'sslPrivateKey' => 'redirect key',
@@ -64,7 +66,7 @@ class SaveSslCertTest extends TestCase
     public function certs_are_skipped_when_not_set(): void
     {
         $project = Project::create(['name' => 'NoSSL Test']);
-        $project->applications()->save($app = new Application([
+        $project->services()->save($service = new ProjectService([
             'domain_name' => 'nossl.test',
             'config' => ['ssl' => true],
         ]));