+++ tests/Feature/Api/Projects/UpdateProjectTest.php
@@ -5,7 +5,6 @@ namespace Tests\Feature\Api\Projects;
 use DMS\PHPUnitExtensions\ArraySubset\ArraySubsetAsserts;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Http\Response;
-use Illuminate\Support\Str;
 use Servidor\Projects\Project;
 use Tests\RequiresAuth;
 use Tests\TestCase;
@@ -58,21 +57,10 @@ class UpdateProjectTest extends TestCase
      *
      * @depends authed_user_can_rename_project
      */
-    public function update_response_includes_applications(array $data): void
+    public function update_response_includes_services(array $data): void
     {
-        $this->assertArrayHasKey('applications', $data);
-        $this->assertEmpty($data['applications']);
-    }
-
-    /**
-     * @test
-     *
-     * @depends authed_user_can_rename_project
-     */
-    public function update_response_includes_redirects(array $data): void
-    {
-        $this->assertArrayHasKey('redirects', $data);
-        $this->assertEmpty($data['redirects']);
+        $this->assertArrayHasKey('services', $data);
+        $this->assertEmpty($data['services']);
     }
 
     /** @test */
@@ -99,10 +87,10 @@ class UpdateProjectTest extends TestCase
      *
      * @dataProvider symlinkProvider
      */
-    public function updating_project_also_toggles_symlink(string $type, array $data): void
+    public function updating_project_also_toggles_symlink(array $data): void
     {
-        $name = $type . ' Symlink Test';
-        $domain = $type . '-symlink.test';
+        $name = 'Service Symlink Test';
+        $domain = 'service-symlink.test';
 
         $this->assertFileDoesNotExist("/etc/nginx/sites-available/{$domain}.conf");
         $this->assertFileDoesNotExist("/etc/nginx/sites-enabled/{$domain}.conf");
@@ -110,7 +98,7 @@ class UpdateProjectTest extends TestCase
         $project->save();
 
         $this->authed()->postJson(
-            '/api/projects/' . $project->id . '/' . Str::plural($type),
+            '/api/projects/' . $project->id . '/services',
             array_merge(['domain' => $domain], $data),
         );
         $this->assertFileExists("/etc/nginx/sites-available/{$domain}.conf");
@@ -129,15 +117,20 @@ class UpdateProjectTest extends TestCase
     public function symlinkProvider(): array
     {
         return [
-            'Project with an application' => ['app', [
+            'Project with a service' => [[
                 'template' => 'html',
-                'provider' => 'github',
-                'repository' => 'dshoreman/servidor-test-site',
-                'branch' => 'develop',
+                'config' => ['source' => [
+                    'provider' => 'github',
+                    'repository' => 'dshoreman/servidor-test-site',
+                    'branch' => 'develop',
+                ]],
             ]],
-            'Project with a redirect' => ['redirect', [
-                'target' => 'example.com',
-                'type' => 301,
+            'Project with a redirect' => [[
+                'template' => 'redirect',
+                'config' => ['redirect' => [
+                    'target' => 'example.com',
+                    'type' => 301,
+                ]],
             ]],
         ];
     }